backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
DateTime==5.3
distro==1.9.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
numpy==1.26.2
openai==1.6.1
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
zope.interface==6.1
